backend/app/servers/poetry_analysis/requirements.txt
sqlalchemy==2.0.35
faiss-cpu==1.9.0
requests==2.32.3
python-dotenv==1.0.1
colorama==0.4.6
fastapi==0.115.2
websockets==13.1
uvicorn==0.32.0
tensorflow==2.17
duckdb==1.1.2
ibm-watsonx-ai==1.1.14
langchain-ibm==0.3.1
langchain-core==0.3.10
langchain-community==0.3.2
langchain-huggingface==0.1.0
langchain-openai
langchain-text-splitters==0.3.0
pandas==2.1.4
tf-keras

